pyproject.toml
[project]
name = "ai-council"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.4.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "ollama>=0.3.0",
    "ddgs>=9.0.0",
    "requests>=2.28.0",
    "google-genai>=1.67.0",
]
